kp_historical_macrodata/requirements.txt
boto3
pandas
kafka-python
requests